Round 110: bug fixes, Siri background mode, UI polish, responsive scaling, safety fixes
- Fix: data written to wrong child in multi-child families (stale resolvedActiveId)
- Fix: quickAddLog infinite retry loop (capped at 10, child ID check)
- Fix: nappy reminder interval dying on every log entry (stale days ref)
- Fix: Firestore merge mutating const (let merged)
- Fix: localStorage persistence race (moved into same updater)
- Fix: cross-midnight nap timer wrong elapsed time
- Fix: widget timer continues after app stop (immediate widget data push)
- Fix: date picker jumping to wrong day (calendar arithmetic)
- Fix: lavender oil tip replaced with NHS safety warning
- Siri: all intents work without opening app (openAppWhenRun=false)
- Siri: multiple commands queue while driving
- Android: deleted duplicate MainActivity.kt
- Android: READ_MEDIA_IMAGES permission for Android 13+
- Android: stopForeground(STOP_FOREGROUND_REMOVE)
- Android: prediction notifications wired up
- Android: PrintPlugin WebView memory leak fixed
- Android: version bump to 2.1.0 (versionCode 21)
- UI: Emil Kowalski polish pass (sheet animations, button feedback, card stagger)
- UI: removed infinite candlePulse from all buttons
- UI: stronger Apple glass effect (blur 24px, sat 1.8)
- UI: responsive scaling system (_rf/_rs helpers)
- UI: nappy log colour chips 7-column grid
- Feature: grandfathered pricing for legacy users
- Feature: restoreChildSyncCodesFromCloud
- Feature: parseTimeFree top-level utility
- Feature: night weaning card in Insights filtered view
- Feature: welcome back card after 2+ day gap
- Feature: review prompt simplified (Love it → store, Needs work → email)
- Feature: night weaning pointer on Today tab

+++ android/app/src/main/java/com/obubba/app/widgets/OBubbaSummaryWidget.java
@@ -114,7 +114,7 @@ public class OBubbaSummaryWidget extends AppWidgetProvider {
                 String lastSide = d.optString("lastBreastSide", "");
                 String side = d.optString("breastSide", "");
                 long startMs = 0;
-                try { if (!d.isNull("timerStartMs")) startMs = (long) d.optDouble("timerStartMs", 0); } catch (Exception x) {}
+                try { if (!d.isNull("timerStartMs")) startMs = d.optLong("timerStartMs", 0); } catch (Exception x) {}
 
                 v.setTextViewText(R.id.tv_baby_name, "\uD83E\uDDF8 " + name);
